requirements.txt
appdirs==1.4.4
asn1crypto==0.24.0
beautifulsoup4==4.6.0
buildozer==0.40.dev0
catfish==1.4.4
certifi==2018.1.18
chardet==3.0.4
colorama==0.4.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
Cython==0.29.10
defer==1.0.6
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
docutils==0.15.2
filelock==3.0.12
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
importlib-metadata==3.3.0
importlib-resources==3.3.0
Jinja2==2.11.2
keyring==10.6.0
keyrings.alt==3.0
Kivy==1.11.1
Kivy-Garden==0.1.4
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lightdm-gtk-greeter-settings==1.2.2
lxml==4.2.1
MarkupSafe==1.1.1
menulibre==2.2.0
more-itertools==7.2.0
mugshot==0.4.0
netifaces==0.10.4
numpy==1.13.3
oauth==1.0.1
olefile==0.45.1
onboard==1.4.1
pep517==0.9.1
pexpect==4.2.1
Pillow==5.1.0
plyer==2.0.0
psutil==5.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pyenchant==2.0.0
Pygments==2.4.2
pygobject==3.26.1
PyOpenGL==3.1.0
PySDL2==0.9.3
python-apt==1.6.5+ubuntu0.4
python-dateutil==2.6.1
python-debian==0.1.32
python-debianbts==2.7.2
python-for-android==2019.8.9
python-magic==0.4.16
pytoml==0.1.21
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scour==0.36
SecretStorage==2.3.1
sgt-launcher==0.2.4
sh==1.14.1
simplejson==3.13.2
six==1.15.0
systemd-python==234
toml==0.10.2
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
virtualenv==20.2.2
wadllib==1.3.2
webencodings==0.5
xcffib==0.5.1
xkit==0.0.0
zipp==3.4.0
zope.interface==4.3.2
